\name{addinfo}
\alias{addinfo}
\title{Add some information}
\usage{
addinfo(raw_counts, gene_data)
}
\arguments{
\item{raw_counts}{Object in R that contains the table of counts generated by Rsubread.}

\item{gene_data}{Object in R that contains gene info (see \link{hg38} dataset as example!).}
}
\description{
It adds extra information to raw_counts generated by RSubread and also eliminates those GeneID whose value in all samples is = 0.
}
\details{
The result is returned in a R object  (raw_counts_addinfo).
}
\examples{
# We prepare the sample data:
c(1,2,10,55,100) -> genes
c(0,5,10,20,50) -> sampleA
c(0,0,15,30,50) -> sampleB
c(0,1,2,3,100) -> sampleC
data.frame("GeneID"= genes,"A_counts"= sampleA,"B_counts"=sampleB,"C_counts"=sampleC) -> test_counts
str(test_counts)
# We apply the function:
addinfo(test_counts,)
# Select the RefSeq file from the Rsubread package.
# WAIT...
raw_counts_addinfo # R object where it is saved
# You can see that the "GeneID" = 1 is missing because all counts in samples are 0.
####################################################################
addinfo(data_RNASeq_raw_counts,) # Internal dataset
# You can see that the "GeneID" = 1,4 & 3000 is missing because all counts in samples are 0.
}
\author{
Jose Jordan
}
