% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_RNASeq_raw_counts.R
\docType{data}
\name{data_RNASeq_raw_counts}
\alias{data_RNASeq_raw_counts}
\title{RNASeq counts data example}
\format{
A tibble: 15 x 6
}
\source{
Own elaboration
}
\usage{
data_RNASeq_raw_counts
}
\description{
RNASeq counts data; example to use in jtools
}
\examples{

}
\keyword{datasets}
