% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{Extract gene length}
\usage{
extract()
}
\description{
Function that extracts the genes and the length of the transcript.
}
\details{
It extracts from the "RefSeq" file of the Rsubread package the genes and length of the transcript and
returns it in an R object (gene_lenght_file).
}
\examples{
hg38 # When you use the function extract(), you can load a gene data file from Rsubread package
}
\author{
Jose Jordan
}
