% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{group}
\alias{group}
\title{Group of cells more informative}
\usage{
group(WantedData, n, Filter_value, num_cells, num_set_seed, num_col)
}
\arguments{
\item{WantedData}{Matrix data. \strong{NECESSARY}}

\item{n}{Number of subsampling. Optional. By default = 100.}

\item{Filter_value}{Cut-off value. Optional. By default = 1.}

\item{num_cells}{Cells to select. Optional. By default = 20 (maximum value allowed. Minimal value allowed is 3).}

\item{num_set_seed}{Set seed value. Optional. By default = 44.}

\item{num_col}{Column number in which the data matrix begins. Optional. By default = 1.}
}
\description{
Group of cells more informative in a dataset
}
\details{
Computes from a table cell counts of RNASeq the group of cells more informative
}
\examples{
no examples yet
}
\author{
Jose Jordan, JuanVi
}
